Added sweetalert to show the notification to user, moderator, admin

+++ Fornt-end/quick-recipe-sharing-platform-ui/src/app/admin/components/recipe/recipe-list/recipe-list.component.ts
@@ -1,4 +1,4 @@
-import { AfterViewInit, ChangeDetectorRef, Component, ElementRef, ViewChild, inject } from '@angular/core';
+import { AfterViewInit, ChangeDetectorRef, Component, ElementRef, OnInit, ViewChild, inject } from '@angular/core';
 import { FormGroup, FormControl } from '@angular/forms';
 import { MatPaginator } from '@angular/material/paginator';
 import { MatTableDataSource } from '@angular/material/table';
@@ -10,6 +10,7 @@ import { RecipeCuisine } from 'src/app/model/recipe-cuisine.model';
 import { CategoryService } from 'src/app/services/category.service';
 import { CuisineService } from 'src/app/services/cuisine.service';
 import { RecipeService } from 'src/app/services/recipe.service';
+import { AlertService } from 'src/app/services/alert.service';
 
 const today = new Date();
 const month = today.getMonth();
@@ -22,27 +23,23 @@ const day = today.getDate();
   templateUrl: './recipe-list.component.html',
   styleUrls: ['./recipe-list.component.scss']
 })
-export class RecipeListComponent implements AfterViewInit {
-  constructor(
-    private recipeService: RecipeService,
-    private changeDetectorRef: ChangeDetectorRef,
-    private categoryService: CategoryService,
-    private cuisineService: CuisineService
-  ) { }
-
-  errorMessage!: string | null;
-  createUserClicked: boolean = false;
-  displayedColumns: string[] = ['S.No', 'Profile', 'Username', 'Title', 'Date Created', 'Deleted At', 'Category', 'Cuisine', 'Actions'];
-  dataSource = new MatTableDataSource<RecipeCuisine>();
-
+export class RecipeListComponent implements AfterViewInit, OnInit {
+  
+  recipeService:RecipeService = inject(RecipeService);
+  changeDetectorRef: ChangeDetectorRef = inject(ChangeDetectorRef);
+  categoryService: CategoryService = inject(CategoryService);
+  cuisineService: CuisineService = inject(CuisineService);
+  alertService: AlertService = inject(AlertService);
   router: Router = inject(Router);
 
+  displayedColumns: string[] = ['S.No', 'Profile', 'Username', 'Title', 'Date Created', 'Deleted At', 'Category', 'Cuisine', 'Actions'];
+  dataSource = new MatTableDataSource<RecipeCuisine>();
   selectedCuisine!: string;
   selectedCategory!: string;
   cuisines!: Cuisine[];
   categories!: Category[];
 
-  campaignOne = new FormGroup({
+  filterDate = new FormGroup({
     start: new FormControl(new Date(year - 1, month, day)),
     end: new FormControl(new Date(year, month, day)),
   });
@@ -53,12 +50,10 @@ export class RecipeListComponent implements AfterViewInit {
   ngOnInit() {
     this.categoryService.getAllCategoryList(0, 100).subscribe((data) => {
       this.categories = data.content;
-      console.log(this.categories);
     });
 
     this.cuisineService.getAllCuisineList(0, 100).subscribe((data) => {
       this.cuisines = data.content;
-      console.log(this.cuisines);
     });
   }
 
@@ -78,10 +73,10 @@ export class RecipeListComponent implements AfterViewInit {
 
   applyFilters() {
     const inputValue = this.messageRef.nativeElement.value;
-    if (!this.selectedCategory && !this.selectedCuisine && !this.campaignOne.value && !inputValue.trim()) {
+    if (!this.selectedCategory && !this.selectedCuisine && !this.filterDate.value && !inputValue.trim()) {
       this.loadPage(this.paginator.pageIndex, this.paginator.pageSize);
     } else {
-      this.makeApiCall(inputValue, this.selectedCuisine, this.selectedCategory, this.campaignOne.value.start, this.campaignOne.value.end);
+      this.makeApiCall(inputValue, this.selectedCuisine, this.selectedCategory, this.filterDate.value.start, this.filterDate.value.end);
     }
   }
 
@@ -92,8 +87,6 @@ export class RecipeListComponent implements AfterViewInit {
     start: Date | null | undefined,
     end: Date | null | undefined
   ) {
-    console.log(start);
-
     this.recipeService.getAllRecipeBySearch(inputValue, selectedCuisine, selectedCategory, start, end, this.paginator.pageIndex, this.paginator.pageSize).subscribe({
       next: (response) => {
         this.dataSource.data = response.content;
@@ -102,10 +95,7 @@ export class RecipeListComponent implements AfterViewInit {
         this.paginator.pageSize = response.size;
       },
       error: (error) => {
-        this.errorMessage = error;
-        setTimeout(() => {
-          this.errorMessage = null;
-        }, 3000);
+        this.alertService.showError("Error in applying the filter");
       },
     });
   }
@@ -119,10 +109,7 @@ export class RecipeListComponent implements AfterViewInit {
         this.paginator.pageSize = response.size;
       },
       error: (error) => {
-        this.errorMessage = error;
-        setTimeout(() => {
-          this.errorMessage = null;
-        }, 3000);
+        this.alertService.showError("Error occurred in loading the recipeList");
       },
     });
   }
@@ -140,32 +127,59 @@ export class RecipeListComponent implements AfterViewInit {
   }
 
   onEditClicked(id: string) {
-    this.router.navigate(['/admin/recipes/update'], { queryParams: { detail: id } });
-  }
-
-  onDeleteClicked(id: string) {
-    this.recipeService.deleteRecipeById(id).subscribe({
-      next: () => {
-        this.router.navigate(['/admin/recipes']);
-      },
-      error: (error) => {
-        this.router.navigate(['/admin/recipes']);
-        this.errorMessage = error;
-        setTimeout(() => {
-          this.errorMessage = null;
-        }, 3000);
-      },
+    this.alertService.confirm("Confirm","Are you sure you want to edit this recipe?").then((confirmed) => {
+      if(confirmed) {
+        this.router.navigate(['/admin/recipes/update'], { queryParams: { detail: id } });
+      }
     });
   }
 
+  onDeleteClicked(id: string, deletedAt: string | null) {
+    if(!deletedAt){
+      this.alertService.confirmDelete("Confirm","Are you sure you want to delete this recipe?").then((confirmed) => {
+        if (confirmed) {
+          this.recipeService.deleteRecipeById(id).subscribe({
+            next: () => {
+              this.alertService.showSuccess("Recipe deleted successfully");
+              window.history.go(-1);
+            },
+            error: (error) => {
+              this.alertService.showError("Error occurred in delete the recipe");
+              window.history.go(-1);
+            },
+          });
+        } else {
+          window.history.go(-1);
+        }
+      });
+    } else {
+      this.alertService.confirmDelete("Confirm","Are you sure you want to revoke this recipe?").then((confirmed) => {
+        if (confirmed) {
+          this.recipeService.deleteRecipeById(id).subscribe({
+            next: () => {
+              this.alertService.showSuccess("Recipe revoked successfully");
+              window.history.go(-1);
+            },
+            error: (error) => {
+              this.alertService.showError("Error occurred in revoke the recipe");
+              window.history.go(-1);
+            },
+          });
+        } else {
+          window.history.go(-1);
+        }
+      });
+    }
+  }
+  
   onStartDateChange(event: MatDatepickerInputEvent<Date>) {
-    this.campaignOne.patchValue({
+    this.filterDate.patchValue({
       start: event.value,
     });
   }
 
   onEndDateChange(event: MatDatepickerInputEvent<Date>) {
-    this.campaignOne.patchValue({
+    this.filterDate.patchValue({
       end: event.value,
     });
   }
